feat: query TFM chains
